backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-multipart==0.0.12
python-dotenv==1.0.1
PyMuPDF==1.24.10
google-generativeai==0.8.3
httpx==0.27.2
beautifulsoup4==4.12.3
